+++ dev-demo/custom/login2.vue
@@ -1,11 +1,14 @@
 <template>
-    <div>
-        setup
+    <div @click="close" class="px-4 py-2 text-sm text-gray-500 dark:text-gray-400 hover:text-gray-600 dark:hover:text-gray-500 hover:underline dark:hover:underline cursor-pointer">
+        🌂
     </div>
 </template>
 
 <script setup>
-
+function close() {
+    console.log('🚀 ~ file: login2.vue ~ line 5 ~ close ~ close');
+    window.adminforth.closeUserMenuDropdown();
+}
 </script>
 
 <style lang="scss" scoped>